Transform correlated no-aggregation query to join
