+++ presto-main/src/main/java/com/facebook/presto/sql/planner/optimizations/TransformCorrelatedScalarAggregationToJoin.java
@@ -22,8 +22,6 @@ import com.facebook.presto.spi.type.BigintType;
 import com.facebook.presto.spi.type.BooleanType;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeSignature;
-import com.facebook.presto.sql.ExpressionUtils;
-import com.facebook.presto.sql.planner.DependencyExtractor;
 import com.facebook.presto.sql.planner.PlanNodeIdAllocator;
 import com.facebook.presto.sql.planner.Symbol;
 import com.facebook.presto.sql.planner.SymbolAllocator;
@@ -31,36 +29,26 @@ import com.facebook.presto.sql.planner.plan.AggregationNode;
 import com.facebook.presto.sql.planner.plan.ApplyNode;
 import com.facebook.presto.sql.planner.plan.AssignUniqueId;
 import com.facebook.presto.sql.planner.plan.EnforceSingleRowNode;
-import com.facebook.presto.sql.planner.plan.FilterNode;
 import com.facebook.presto.sql.planner.plan.JoinNode;
-import com.facebook.presto.sql.planner.plan.LimitNode;
 import com.facebook.presto.sql.planner.plan.PlanNode;
 import com.facebook.presto.sql.planner.plan.ProjectNode;
 import com.facebook.presto.sql.planner.plan.SimplePlanRewriter;
-import com.facebook.presto.sql.tree.DefaultTraversalVisitor;
 import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.FunctionCall;
-import com.facebook.presto.sql.tree.LogicalBinaryExpression;
 import com.facebook.presto.sql.tree.QualifiedName;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Iterables;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 import static com.facebook.presto.sql.analyzer.TypeSignatureProvider.fromTypeSignatures;
 import static com.facebook.presto.sql.planner.optimizations.PlanNodeSearcher.searchFrom;
 import static com.facebook.presto.sql.planner.optimizations.Predicates.isInstanceOfAny;
 import static com.facebook.presto.sql.planner.plan.SimplePlanRewriter.rewriteWith;
 import static com.facebook.presto.sql.tree.BooleanLiteral.TRUE_LITERAL;
-import static com.facebook.presto.util.ImmutableCollectors.toImmutableList;
 import static com.facebook.presto.util.ImmutableCollectors.toImmutableMap;
 import static java.util.Objects.requireNonNull;
 
@@ -119,12 +107,14 @@ public class TransformCorrelatedScalarAggregationToJoin
         private final PlanNodeIdAllocator idAllocator;
         private final SymbolAllocator symbolAllocator;
         private final Metadata metadata;
+        private final PlanNodeDecorrelator decorrelator;
 
         public Rewriter(PlanNodeIdAllocator idAllocator, SymbolAllocator symbolAllocator, Metadata metadata)
         {
             this.idAllocator = requireNonNull(idAllocator, "idAllocator is null");
             this.symbolAllocator = requireNonNull(symbolAllocator, "symbolAllocator is null");
             this.metadata = requireNonNull(metadata, "metadata is null");
+            this.decorrelator = new PlanNodeDecorrelator(idAllocator);
         }
 
         @Override
@@ -146,7 +136,7 @@ public class TransformCorrelatedScalarAggregationToJoin
         private PlanNode rewriteScalarAggregation(ApplyNode apply, AggregationNode aggregation)
         {
             List<Symbol> correlation = apply.getCorrelation();
-            Optional<DecorrelatedNode> source = decorrelateFilters(aggregation.getSource(), correlation);
+            Optional<PlanNodeDecorrelator.DecorrelatedNode> source = decorrelator.decorrelateFilters(aggregation.getSource(), correlation);
             if (!source.isPresent()) {
                 return apply;
             }
@@ -259,171 +249,6 @@ public class TransformCorrelatedScalarAggregationToJoin
                     scalarAggregation.getStep(),
                     scalarAggregation.getHashSymbol(),
                     Optional.empty()));
-        }
-
-        private Optional<DecorrelatedNode> decorrelateFilters(PlanNode node, List<Symbol> correlation)
-        {
-            PlanNodeSearcher filterNodeSearcher = searchFrom(node)
-                    .where(FilterNode.class::isInstance)
-                    .skipOnlyWhen(isInstanceOfAny(ProjectNode.class, LimitNode.class));
-            List<FilterNode> filterNodes = filterNodeSearcher.findAll();
-
-            if (filterNodes.isEmpty()) {
-                return decorrelatedNode(ImmutableList.of(), node, correlation);
-            }
-
-            if (filterNodes.size() > 1) {
-                return Optional.empty();
-            }
-
-            FilterNode filterNode = filterNodes.get(0);
-            Expression predicate = filterNode.getPredicate();
-
-            if (!isSupportedPredicate(predicate)) {
-                return Optional.empty();
-            }
-
-            if (!DependencyExtractor.extractUnique(predicate).containsAll(correlation)) {
-                return Optional.empty();
-            }
-
-            Map<Boolean, List<Expression>> predicates = ExpressionUtils.extractConjuncts(predicate).stream()
-                    .collect(Collectors.partitioningBy(isUsingPredicate(correlation)));
-            List<Expression> correlatedPredicates = ImmutableList.copyOf(predicates.get(true));
-            List<Expression> uncorrelatedPredicates = ImmutableList.copyOf(predicates.get(false));
-
-            node = updateFilterNode(filterNodeSearcher, uncorrelatedPredicates);
-
-            if (!correlatedPredicates.isEmpty()) {
-                // filterNodes condition has changed so Limit node no longer applies for EXISTS subquery
-                node = removeLimitNode(node);
-            }
-
-            node = ensureJoinSymbolsAreReturned(node, correlatedPredicates);
-
-            return decorrelatedNode(correlatedPredicates, node, correlation);
-        }
-
-        private static Optional<DecorrelatedNode> decorrelatedNode(
-                List<Expression> correlatedPredicates,
-                PlanNode node,
-                List<Symbol> correlation)
-        {
-            if (DependencyExtractor.extractUnique(node).stream().anyMatch(correlation::contains)) {
-                // node is still correlated ; /
-                return Optional.empty();
-            }
-            return Optional.of(new DecorrelatedNode(correlatedPredicates, node));
-        }
-
-        private static Predicate<Expression> isUsingPredicate(List<Symbol> symbols)
-        {
-            return expression -> symbols.stream().anyMatch(DependencyExtractor.extractUnique(expression)::contains);
-        }
-
-        private PlanNode updateFilterNode(PlanNodeSearcher filterNodeSearcher, List<Expression> newPredicates)
-        {
-            if (newPredicates.isEmpty()) {
-                return filterNodeSearcher.removeAll();
-            }
-            FilterNode oldFilterNode = Iterables.getOnlyElement(filterNodeSearcher.findAll());
-            FilterNode newFilterNode = new FilterNode(
-                    idAllocator.getNextId(),
-                    oldFilterNode.getSource(),
-                    ExpressionUtils.combineConjuncts(newPredicates));
-            return filterNodeSearcher.replaceAll(newFilterNode);
-        }
-
-        private static PlanNode removeLimitNode(PlanNode node)
-        {
-            node = searchFrom(node)
-                    .where(LimitNode.class::isInstance)
-                    .skipOnlyWhen(ProjectNode.class::isInstance)
-                    .removeFirst();
-            return node;
-        }
-
-        private PlanNode ensureJoinSymbolsAreReturned(PlanNode scalarAggregationSource, List<Expression> joinPredicate)
-        {
-            Set<Symbol> joinExpressionSymbols = DependencyExtractor.extractUnique(joinPredicate);
-            ExtendProjectionRewriter extendProjectionRewriter = new ExtendProjectionRewriter(
-                    idAllocator,
-                    joinExpressionSymbols);
-            return rewriteWith(extendProjectionRewriter, scalarAggregationSource);
-        }
-
-        private static boolean isSupportedPredicate(Expression predicate)
-        {
-            AtomicBoolean isSupported = new AtomicBoolean(true);
-            new DefaultTraversalVisitor<Void, AtomicBoolean>()
-            {
-                @Override
-                protected Void visitLogicalBinaryExpression(LogicalBinaryExpression node, AtomicBoolean context)
-                {
-                    if (node.getType() != LogicalBinaryExpression.Type.AND) {
-                        context.set(false);
-                    }
-                    return null;
-                }
-            }.process(predicate, isSupported);
-            return isSupported.get();
-        }
-    }
-
-    private static class DecorrelatedNode
-    {
-        private final List<Expression> correlatedPredicates;
-        private final PlanNode node;
-
-        public DecorrelatedNode(List<Expression> correlatedPredicates, PlanNode node)
-        {
-            requireNonNull(correlatedPredicates, "correlatedPredicates is null");
-            this.correlatedPredicates = ImmutableList.copyOf(correlatedPredicates);
-            this.node = requireNonNull(node, "node is null");
-        }
-
-        Optional<Expression> getCorrelatedPredicates()
-        {
-            if (correlatedPredicates.isEmpty()) {
-                return Optional.empty();
-            }
-            return Optional.of(ExpressionUtils.and(correlatedPredicates));
-        }
-
-        public PlanNode getNode()
-        {
-            return node;
-        }
-    }
-
-    private static class ExtendProjectionRewriter
-            extends SimplePlanRewriter<PlanNode>
-    {
-        private final PlanNodeIdAllocator idAllocator;
-        private final Set<Symbol> symbols;
-
-        ExtendProjectionRewriter(PlanNodeIdAllocator idAllocator, Set<Symbol> symbols)
-        {
-            this.idAllocator = requireNonNull(idAllocator, "idAllocator is null");
-            this.symbols = requireNonNull(symbols, "symbols is null");
-        }
-
-        @Override
-        public PlanNode visitProject(ProjectNode node, RewriteContext<PlanNode> context)
-        {
-            ProjectNode rewrittenNode = (ProjectNode) context.defaultRewrite(node, context.get());
-
-            List<Symbol> symbolsToAdd = symbols.stream()
-                    .filter(rewrittenNode.getSource().getOutputSymbols()::contains)
-                    .filter(symbol -> !rewrittenNode.getOutputSymbols().contains(symbol))
-                    .collect(toImmutableList());
-
-            Map<Symbol, Expression> assignments = ImmutableMap.<Symbol, Expression>builder()
-                    .putAll(rewrittenNode.getAssignments())
-                    .putAll(toAssignments(symbolsToAdd))
-                    .build();
-
-            return new ProjectNode(idAllocator.getNextId(), rewrittenNode.getSource(), assignments);
         }
     }
